removed assets directory
